+++ viewer_2.0/globals.py
@@ -1,14 +1,17 @@
-import numpy as np
-from reuss import config as cfg
-import multiprocessing as mp
 import ctypes
+import numpy as np
+import multiprocessing as mp
 
+from reuss import config as cfg
+
+stream = "tcp://localhost:4545"
 
 #Configuration
 nrow = cfg.nrows()
 ncol = cfg.ncols()
 
 dtype = np.float32
+cdtype = ctypes.c_float
 
 accframes = 0
 acc_image = np.zeros((nrow,ncol), dtype = dtype)
