+++ viewer_2.0/overlay.py
@@ -2,9 +2,10 @@ import json
 import logging
 import numpy as np
 import pyqtgraph as pg
-from reuss import config as cfg
 from PySide6.QtWidgets import QGraphicsEllipseItem, QGraphicsRectItem
 from PySide6.QtCore import QRectF
+
+from reuss import config as cfg
 
 def draw_overlay(image_item):
     for key, value in cfg.parser['overlay'].items():
@@ -13,7 +14,8 @@ def draw_overlay(image_item):
         if 'circle' in key:
             x, y = s['xy']
             r = s['radius']
-            # QGraphicsEllipseItem takes QRectF as argument, which defines the bounding rectangle.
+            # QGraphicsEllipseItem takes QRectF as argument, 
+            # which defines the bounding rectangle.
             # QRectF(x, y, width, height)
             item = QGraphicsEllipseItem(QRectF(x-r, y-r, 2*r, 2*r))
         elif 'rectangle' in key:
