+++ viewer_2.0/tem_controls/fit_beam_intensity.py
@@ -1,12 +1,13 @@
 import logging
-from pathlib import Path
 import numpy as np
-from scipy.interpolate import griddata
-from lmfit import Model, Parameters
-import globals
 import pyqtgraph as pg
+from pathlib import Path
 from PySide6 import QtWidgets
+from lmfit import Model, Parameters
+from scipy.interpolate import griddata
 from line_profiler import LineProfiler
+
+import globals
 
 # Define a rotated 2D Gaussian function
 def gaussian2d_rotated(x, y, amplitude, xo, yo, sigma_x, sigma_y, theta):
