+++ viewer_2.0/toggle_button.py
@@ -1,0 +1,6 @@
+from PySide6.QtWidgets import QPushButton
+
+class ToggleButton(QPushButton):
+    def __init__(self, label, window):
+        super().__init__(label, window)
+        self.started = False
